Adds: Business interface Fixes: onCreate Override in MainActivity
