+++ OrcaStars/app/src/main/java/edu/cascadia/orcastars/DB/Business.java
@@ -1,0 +1,12 @@
+package edu.cascadia.orcastars.DB;
+
+public interface Business {
+        int getId();
+        String getName();
+        String getDescription();
+        String getHours();
+        String getLocation();
+        int getLogo();
+        int getPhoto();
+
+}
