+++ OrcaStars/app/src/main/java/edu/cascadia/orcastars/OrcaStarsApp.java
@@ -6,10 +6,8 @@ import edu.cascadia.orcastars.DB.AppDatabase;
 
 public class OrcaStarsApp extends Application {
     private AppExecutors mAppExecutors;
-    /*@Override*/
-    /*Uncommented for now, because AppExecutors does not actually
-    have an onCreate to override*/
-    public void OnCreate(){
+    @Override
+    public void onCreate(){
         super.onCreate();
         mAppExecutors = new AppExecutors();
     }
